moved to signed idx_t

+++ Source/GPU/Vulkan/States/General/General_Pipeline.hpp
@@ -35,7 +35,7 @@ struct General_Pipeline
             .sType                      = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO,
             .pNext                      = nullptr,
             .flags                      = 0,
-            .stageCount                 = ArrayCount(shader.stageInfo),
+            .stageCount                 = (uint32_t) ArrayCount(shader.stageInfo),
             .pStages                    = shader.stageInfo,
             .pVertexInputState          = &vertexInput,
             .pInputAssemblyState        = &inputAssembly,
